# api/docs/openapi.yaml
openapi: '3.0.0'
...
servers: []
info:
  version: '1.0'
  title: Guarde API
  description: 'API for mobile application and platform used in team 38 solution'
  contact:
    name: Diego De Santiago
    url: 'https://github.com/diegodisant'
    email: diegodisan@protonmail.com
  license:
    name: MIT
    url: 'https://github.com/diegodisant/team38-platform'
paths:
...
  /quarantined/{mobileUserId}:
    put:
      summary: Quarantined change
      description: Marks if the user needs to be quarantined
      tags:
        - mobile
      operationId: changeQuarantinedStatus
      parameters:
        - in: path
          name: mobileUserId
          schema:
            type: string
          required: true
          description: Mongo unique user identifer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuarantinedRequest'
      responses:
        '201':
          description: User marked as quarantined
          content:
            application/json:
              schema:
                properties:
                  statusCode:
                    type: number
                    description: Server status code
                  message:
                    type: string
                    description: Server message response
                example:
                  statusCode: 201
                  message: 'Mobile user marked as quarantined'
        '406':
          description: Schema Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: Database error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseErrorResponse'
    get:
      summary: Quarantined check
      description: Checks if the user is quarantined
      tags:
        - mobile
      operationId: getQuarantinedStatus
      parameters:
        - in: path
          name: mobileUserId
          schema:
            type: string
          required: true
          description: Mongo unique user identifer
      responses:
        '200':
          description: Quarantined user status
          content:
            application/json:
              schema:
                properties:
                  isQuarantined:
                    type: boolean
                    description: Quarantined user status
                example:
                  isQuarantined: true
        '406':
          description: Schema Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: Database error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseErrorResponse'
  /lab-analysis:
    post:
      summary: Laboratory analysis register
      description: Register a laboratory analysis result
      tags:
        - api
      operationId: createLabAnalysis
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabAnalysisRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabAnalysisResponse'
        '406':
          description: Schema Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: Database error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseErrorResponse'
    get:
      summary: All laboratory analysis
      description: Retrieves all laboratory analysis
      tags:
        - api
      operationId: getAllLaboratoryAnalysis
      responses:
        '200':
          description: All laboratory analysis
          content:
            application/json:
              schema:
                properties:
                  lab-analysis:
                    type: array
                    items:
                      $ref: '#/components/schemas/LabAnalysisResponse'
        '406':
          description: Schema Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: Database error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseErrorResponse'
...
components:
  schemas:
...
    QuarantinedRequest:
      $ref: 'components/Quarantined/QuarantinedRequest.yaml#/QuarantinedRequest'
    LabAnalysisRequest:
      $ref: 'components/LabAnalysis/LabAnalysisRequest.yaml#/LabAnalysisRequest'
    LabAnalysisResponse:
      $ref: 'components/LabAnalysis/LabAnalysisResponse.yaml#/LabAnalysisResponse'
...
    DatabaseErrorResponse:
      $ref: 'components/Error/DatabaseErrorResponse.yaml#/DatabaseErrorResponse'
    ValidationErrorResponse:
      $ref: 'components/Error/ValidationErrorResponse.yaml#/ValidationErrorResponse'